# transport: refactor to reduce lock contention and improve performance